+++ src/main/java/com/karankumar/bookproject/ui/settings/SettingsView.java
@@ -19,16 +19,9 @@ package com.karankumar.bookproject.ui.settings;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.karankumar.bookproject.backend.dto.GoodreadsBookImport;
-import com.karankumar.bookproject.backend.entity.Author;
-import com.karankumar.bookproject.backend.entity.Book;
-import com.karankumar.bookproject.backend.entity.CustomShelf;
-import com.karankumar.bookproject.backend.entity.PredefinedShelf;
-import com.karankumar.bookproject.backend.entity.RatingScale;
 import com.karankumar.bookproject.backend.service.BookService;
-import com.karankumar.bookproject.backend.service.CustomShelfService;
-import com.karankumar.bookproject.backend.service.PredefinedShelfService;
+import com.karankumar.bookproject.backend.service.ImportService;
 import com.karankumar.bookproject.backend.utils.CsvUtils;
-import com.karankumar.bookproject.backend.util.PredefinedShelfUtils;
 import com.karankumar.bookproject.ui.MainView;
 import com.karankumar.bookproject.ui.components.dialog.ResetShelvesDialog;
 import com.karankumar.bookproject.ui.components.toggle.SwitchToggle;
@@ -51,19 +44,11 @@ import com.vaadin.flow.server.StreamRegistration;
 import com.vaadin.flow.server.StreamResource;
 import com.vaadin.flow.theme.lumo.Lumo;
 import lombok.extern.java.Log;
-import org.apache.commons.lang3.StringUtils;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
-import java.time.LocalDate;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.function.Predicate;
 import java.util.logging.Level;
-import java.util.stream.Collectors;
 
 @Route(value = "settings", layout = MainView.class)
 @PageTitle("Settings | Book Project")
@@ -96,8 +81,7 @@ public class SettingsView extends HorizontalLayout {
     private static final Upload importGoodreadsUpload = new Upload(importGoodreadsMemoryBuffer);
 
     private final BookService bookService;
-    private final transient PredefinedShelfService predefinedShelfService;
-    private final transient CustomShelfService customShelfService;
+    private final transient ImportService importService;
 
     static {
         configureDarkModeToggle();
@@ -135,11 +119,9 @@ public class SettingsView extends HorizontalLayout {
         importGoodreadsUpload.setUploadButton(new Button(IMPORT_FROM_GOODREADS));
     }
 
-    SettingsView(BookService bookService, PredefinedShelfService predefinedShelfService,
-                 CustomShelfService customShelfService) {
+    SettingsView(BookService bookService, ImportService importService) {
         this.bookService = bookService;
-        this.predefinedShelfService = predefinedShelfService;
-        this.customShelfService = customShelfService;
+        this.importService = importService;
 
         setDarkModeState();
 
@@ -200,92 +182,11 @@ public class SettingsView extends HorizontalLayout {
                                 GoodreadsBookImport.class);
                 goodreadsBookImports.forEach(i -> LOGGER.info(i.toString()));
 
-                saveAll(goodreadsBookImports);
+                importService.importGoodreadsBooks(goodreadsBookImports);
             } catch (IOException e) {
                 LOGGER.severe("Error in reading input file, error: + " + e);
             }
         });
-    }
-
-    private void saveAll(List<GoodreadsBookImport> goodreadsBookImports) {
-        List<Book> books = toBooks(goodreadsBookImports);
-        List<Book> savedBooks = books.stream()
-                                     .map(bookService::save)
-                                     .filter(Optional::isPresent)
-                                     .map(Optional::get)
-                                     .collect(Collectors.toList());
-        savedBooks.forEach(b -> LOGGER.info("Book: " + b + " saved successfully"));
-    }
-
-    private List<Book> toBooks(Collection<? extends GoodreadsBookImport> goodreadsBookImports) {
-        return goodreadsBookImports.stream()
-                                   .map(this::toBook)
-                                   .filter(Optional::isPresent)
-                                   .map(Optional::get)
-                                   .collect(Collectors.toList());
-    }
-
-    private Optional<Book> toBook(GoodreadsBookImport goodreadsBookImport) {
-        Optional<Author> author = toAuthor(goodreadsBookImport.getAuthor());
-        if (author.isEmpty()) {
-            LOGGER.severe("Author is null");
-            return Optional.empty();
-        }
-
-        Optional<PredefinedShelf> predefinedShelf =
-                toPredefinedShelf(goodreadsBookImport.getBookshelves(),
-                        goodreadsBookImport.getDateRead());
-        if (predefinedShelf.isEmpty()) {
-            LOGGER.severe("Predefined shelf is null");
-            return Optional.empty();
-        }
-
-        Book book = new Book(goodreadsBookImport.getTitle(), author.get(), predefinedShelf.get());
-
-        Optional<CustomShelf> customShelf = toCustomShelf(goodreadsBookImport.getBookshelves());
-        customShelf.ifPresent(book::setCustomShelf);
-
-        if (Objects.nonNull(goodreadsBookImport.getRating())) {
-            Optional<RatingScale> ratingScale =
-                    RatingScale.of(goodreadsBookImport.getRating() * 2);
-            ratingScale.ifPresent(book::setRating);
-        }
-
-        return Optional.of(book);
-    }
-
-    private Optional<Author> toAuthor(String name) {
-        if (StringUtils.isBlank(name)) {
-            return Optional.empty();
-        }
-        String[] authorNames = name.split(" ");
-        return Optional
-                .of(new Author(authorNames[0], authorNames.length > 1 ? authorNames[1] : null));
-    }
-
-    private Optional<PredefinedShelf> toPredefinedShelf(String shelves, LocalDate dateRead) {
-        if (Objects.nonNull(dateRead)) {
-            return Optional.of(predefinedShelfService.findToReadShelf());
-        }
-        if (StringUtils.isBlank(shelves)) {
-            return Optional.empty();
-        }
-        String[] shelvesArray = shelves.split(",");
-        return Arrays.stream(shelvesArray)
-                     .filter(PredefinedShelfUtils::isPredefinedShelf)
-                     .findFirst()
-                     .map(predefinedShelfService::getPredefinedShelfByNameAsString);
-    }
-
-    private Optional<CustomShelf> toCustomShelf(String shelves) {
-        if (StringUtils.isBlank(shelves)) {
-            return Optional.empty();
-        }
-        String[] shelvesArray = shelves.split(",");
-        return Arrays.stream(shelvesArray)
-                     .filter(Predicate.not(PredefinedShelfUtils::isPredefinedShelf))
-                     .findFirst()
-                     .map(customShelfService::findOrCreate);
     }
 
     private String generateJsonResource() {
